FYP-2/flask-server/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyMuPDF==1.23.8
sentence-transformers==2.2.2
scikit-learn==1.3.2
google-generativeai==0.3.2
reportlab==4.0.4
python-dotenv==1.0.0
torch==2.1.0
numpy==1.24.3
transformers==4.35.2
gunicorn==21.2.0
